removes unused routes/replaces newCase
