+++ api/app.js
@@ -25,8 +25,7 @@ const profileRouter = require('./profile/profileRouter');
 const dsRouter = require('./dsService/dsRouter');
 const judgeRouter = require('./judges/judgeRouter');
 const caseRouter = require('./cases/caseRouter');
-const manageRouter = require('./manageCases/manageRouter');
-const newCaseRouter = require('./newCases/newCaseRouter');
+const pendingCaseRouter = require('./pendingCases/pendingCaseRouter');
 const tagRouter = require('./tags/tagRouter');
 const uploadRouter = require('./upload/uploadRouter');
 const faqRouter = require('./faq/faqRouter');
@@ -65,8 +64,7 @@ app.use('/upload', uploadRouter);
 app.use(['/profile', '/profiles'], profileRouter);
 app.use(['/judge', '/judges'], judgeRouter);
 app.use(['/case', '/cases'], caseRouter);
-app.use('/manage', manageRouter);
-app.use('/newcase', newCaseRouter);
+app.use('/pendingCase', pendingCaseRouter);
 app.use('/tags', tagRouter);
 app.use('/data', dsRouter);
 app.use('/faq', faqRouter);
